requirements.txt
asgiref==3.8.1
Flask==3.1.0
Flask-Cors==5.0.0
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-testing==1.70.0
grpcio-tools==1.70.0
gunicorn==23.0.0
jwt==1.3.1
mysqlclient==2.2.7
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
peewee==3.17.8
phonenumbers==8.13.53
pycryptodome==3.21.0
pytest==8.3.4
requests==2.32.3
smswithoutborders_libsig @ git+https://github.com/smswithoutborders/lib_signal_double_ratchet_python.git@v0.1.2
SwobThirdPartyPlatforms @ git+https://github.com/smswithoutborders/SMSWithoutBorders-Customized-Third-Party-Platforms.git@main#egg=SwobThirdPartyPlatforms
tqdm==4.67.1
twilio==9.4.3
